Be opinionated about DST when sorting timezones
On September 8, 2013 at about 5:44 AM IST (GMT+0530), the
sorted_timezones function failed with
pytz.exceptions.NonExistentTimeError: 2013-09-08 00:14:11.363049.

This appears to be related to an ongoing DST transition in
America/Santiago and setting is_dst=False fixes that. However,
the is_dst parameter doesn't work for the UTC zone, so that's
skipped and added back later.