Fix compatibility with psycopg3 and Python 3.11's tomllib (#374)
Unlike psycopg2, version 3 will not cast a string to an int when querying against an integer column. `IdMixin.url_id` now supplies a comparator that does this, but a query against `url_id` in `BaseScopedIdMixin` (which is a column without a custom comparator) must explicitly be cast to an int (as used in one test for the `load_models` decorator).